thinking on how user settings will look
